link: fix qemu entrypoint as _start
qemu actually jumps to 0x80000000, not the entrypoint, so if we added
M-mode code to another file by putting it into .text.init, we might
find that the change in order leads to that being the entrypoint
instead of _start -- which is especially devious when that code is
sometimes optimized out.

Put _start in its own section, .text.entrypoint, and ensure that this
new section is always ordered before the rest of .text.init.

Also, fix my incorrect notes about this.