uart: guess initial config based on vendor ID
It's important that we can use the UART before initialization has
completed, and so before we know the exact device from the device tree.

Previously, we hardcoded this for all platforms, and if we needed to
have multiple platforms, we switched the default to match whichever one
we were using. Instead of this, guess at the platform based on the
vendor ID, so that it's more likely we get the correct configuration
without needing to explicitly update it for each platform.